(Depends on D4070) Incorporated IntermediateStore features into IntermediateStorage, fixed tests and subclasses accordingly.
Summary: IntermediateStore acts as a layer of indirection between IntermediateStorage and ObjectStore, but it more obfuscates how they are related than helps. This change aims to deprecate and remove IntermediateStore entirely by moving its functionality a layer up, to the IntermediateStorage layer.

Test Plan: Modified IntermediateStore tests to instead test the new IntermediateStorage functionality. Propagated the removal of IntermediateStore to a ton of tests.

Reviewers: sandyryza, nate

Reviewed By: sandyryza

Subscribers: sandyryza

Differential Revision: https://dagster.phacility.com/D4098