Initialize docker grpc server with target
Summary: This diff updates the image code for dagster-core-docker-buildkite to match the code in the `api_tests_repo.py` file. It also updates the `docker-compose` file to initialize the GRPC server with a target, so that we can start adding tests against the repo in the image. The first test added here is for `list_repositories`

Test Plan: unit

Reviewers: nate, aj.nadel, dgibson

Reviewed By: nate

Differential Revision: https://dagster.phacility.com/D3987