(instance-scheduler-7) Add ability to turn off  unloadedable schedules with warning
Summary: This diff adds the ability to turn off schedules for which we don't have the schedule definition loaded. The caveat is that users can't turn these schedules back on. To help protect users against accidentally turning off these schedules, there is a small confirmation window that this is shown before turning the schedule off.

Test Plan:
In this video, we turn on schedules in workspace A, switch to workspace B (without the schedule definitions), and turn off the schedules after clicking through the confirmation.

{F231260}

Reviewers: prha, alangenfeld, aj.nadel

Reviewed By: prha

Subscribers: bengotow

Differential Revision: https://dagster.phacility.com/D4120