Add more out-of-process execution coverage to dagster-graphql tests
Summary:
Right now the only test coverage we have on pipeline launches is using an in-process repo location, whereas our users are all using out-of-process. Increase the test matrix here so that we are testing cli api / grpc execution as well.

(as part of this, rename the grpc environments to 'managed_grpc' and 'deployed_grpc' to try to make the difference between them clearer.

Test Plan: BK

Reviewers: alangenfeld, max, schrockn

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D4225