Fix grpc server heartbeat thread
Summary:
float is a value type, so we can't pass the time to the thread as an arg and expect it to update when the thread ones.

Based on my incomplete understanding of Python thread safety this should be thread-safe I think? Or should I add a lock around self.__list_heartbeat_time??

Test Plan: BK test that would be failing before this change, run dagit, no more server shutdown after 30 seconds

Reviewers: alangenfeld, max

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D4235