deprecate pipeline_run config in helm values
Summary:
This diff attempts to clarify the responsibilities of the "Pipeline Run" section of the Helm values file to be the image for k8s jobs that the celery worker launches, only when user deployments are not enabled. As such, I removed the constraint that the image is required. Additionally, i think env_config_maps and env_secrets fields in "Pipeline Run" section were not actually used so removed those those too.

We also update the scheduler config to allow specifying a separate image to detangle scheduler from "Pipeline Run" image. Recommended image for the scheduler is probably "dagster-fixed" since it only needs to launch the scheduled execution and thus does not need access to user code.

Test Plan: bk

Reviewers: johann, nate

Reviewed By: johann

Differential Revision: https://dagster.phacility.com/D4281