Store start time in UTC when a schedule starts
Summary: This is to give the new non-cron scheduler (which fills in any missing scheduled executions every times it runs) a good starting point. The first time you run the scheduler, it will use the time that you actually pressed start as a starting point.

Test Plan: New BK coverage

Reviewers: alangenfeld, sashank, prha, sandyryza

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D4350