Move Dask DataFrame read/to options under read/to keys (#2821)
* Place existing DataFrame loader configs under a "read" key

The DataFrame loader configs currently specify where to load data from.
Place these under a key named “read to make room for additional loader
options.

* Rename file_type/file_options to read_type/read_options

* Place existing DataFrame materializer configs under a "to" key

The DataFrame materializer configs currently specify where to write
data to. Place these under a key named “to” to make room for additional
materializer options.

* Rename file_type/file_options to to_type/to_options

* Allow materialization to multiple formats

Modify the DataFrame materializer to allow materialization in multiple
formats.

* Move materializer after loader

* Fix capitalization of key option

* Fix spelling of storage_options option

* Generate DataFrame loader config

Separate the data for the DataFrame loader into a dict and create a
function that generates the loader config using it. This reduces
repetition and will be used to provide both the original form of the
loader config and the new form with the read key.

The descriptions of the read options are also cut down to the core
sentence instead of being a copy of the full Dask documentation.

* Read dataframe using function in types

Replace the long if/else list for each read function with a lookup to
DataFrameReadTypes for metadata about the read type, which includes the
function and whether it is path based. Use this to prepare the
arguments to the specified read function, then call it and return the
resulting Dask dataframe.

* Add support for read types at root of loader config

Temporarily add support for specify a read type (e.g.m parquet) at the
root of a loader config. This form is deprecated in favour of the read
key.

* Generate DataFrame materializer config

Separate the data for the DataFrame materializer into a dict and create
a function that generates the materializer config using it. This
reduces repetition and will be used to provide both the original form
of the materializer config and the new form with the read key.

The descriptions of the to options are also cut down to the core
sentence instead of being a copy of the full Dask documentation.

* Write dataframe using function in types

Replace the long if/else list for each to function with a lookup to
DataFrameToTypes for metadata about the to type, which includes the
function and whether it is path based. Use this to prepare the
arguments to the specified to function, then call it.

Also, limit the AssetMaterialization yield to just file based
assets for this commit.

* Consider to specs on config root

Accept to specifications on the root of the type config in addition to
those under the to key. This is temporary for compatibility purposes,
and should be considered deprecated.

* Remove dict_without_keys

The dict_without_keys function is not used.

* Add filters field to read_parquet

This field was accidentally lost in the conversion.

* Move input tests before output tests

Order the tests from input to output.

* Test that results of old and new form reads are the same

* Allow permissive fields on read/to options

Allow extra fields to be specified when loading or materializing
dataframes. The extra options are passed to the underlying Pandas
implementations.

* Fix string quote issues from conversion to double quotes

* Warn on use of read/to configs without read/to keys

Dataframe read/to configs should be placed under read/to keys, not at
the root of the type config.

* Test for warnings on lack of read/to keys

* Change config options to Selector

Change the root level config for both loader and materializer from
Shape to Selector. This will change back to Shape when additional
options are introduced later on.

* Note deprecated config options

See https://github.com/dagster-io/dagster/issues/2872 for information
about deprecated loader and materializer config options.

* Adjust code formatting based on Black