#2887 enable subset reexecution when part of selection failed
Summary:
https://github.com/dagster-io/dagster/issues/2887

current: every step in the selection has to be either finished or succeeded

for example, in pipeline like {F278318}
currently in dagit, when branch_1_0 failed and branch_2_* is still running, the user can't re-execute `branch_1_0*` because branch_1_1 hasn't started.
{F278319}

so we want to enable subset re-execution for selection where part of the selected steps failed and pipeline is still running.

Test Plan: {F278309}

Reviewers: max, catherinewu, prha

Reviewed By: prha

Subscribers: schrockn

Differential Revision: https://dagster.phacility.com/D4384