[dagit] Don't attempt PipelineExecutionRootQuery if no repository
Summary:
Resolves #2953.

In cases where we have no repository available when loading a Playground, there is a brief error because we can't execute a query without the repository or repository name.

To resolve this, first use the `loading` state of the repository query to determine whether the Playground query should be executed at all. In the loading state, show a spinner.

 If we're no longer loading but have no repository information, it's an empty workspace and we can just show the `NonIdealState`.

Test Plan: View Playground tab on cold load, verify that the GraphQL error does not occur and that the Playground tab loads correctly. Repeat with an empty workspace, verify that the empty state appears correctly.

Reviewers: dgibson, alangenfeld

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D4532