[dagit] Display operation name in GraphQL error dialog
Summary:
Resolves #2954

While investigating #2953, I found that I didn't have simple access to the operation name of the query that was erroring.

This diff adds it to the error dialog, which will hopefully help with debugging.

Test Plan: Force GraphQL error as in #2953. Open debug dialog from toast, verify that the operation name is present.

Reviewers: dgibson, alangenfeld, sashank

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D4541