[active execution] throw on lost steps
Summary:
Currently we can end a pipeline in success despite losing communication with step processes since step failure events were not observed and no exception was thrown.

This diff turns `ActiveExecution` in to a context manager, then throws on `__exit__` if the plan is not complete or if steps were in an unknown state

depends on D4591

Test Plan:
Need to think on how to get this in a pytest.
Manually commented out the step failure event in the multiprocess executor and used `error_monster` with a process crash to simulate lost step comms.
{F305789}

Reviewers: dgibson, catherinewu

Reviewed By: dgibson

Differential Revision: https://dagster.phacility.com/D4593