[dagit] show python error cause in log row
Summary:
The way things currently work is that the row for an error doesnt have the `expand` action if it doesn't overflow. This means that if there is a `cause` that we would show in the modal but the error is to short to overflow the log row, there is no way to get to the modal to see the exception.

This includes a `cause` section in the log row so that its is visible and more likley to overflow and make the modal available.

There might be a better systemic fix.

Test Plan:
manual with updated error monster
[screenshot]

Reviewers: dish, bengotow, sashank

Reviewed By: sashank

Differential Revision: https://dagster.phacility.com/D4613