Add utility options to Dask DataFrame type loader/materializer (#2888)
* Create function for sanitizing column names

Create a utility function for sanitizing column names in a DataFrame.
Currently, this just lowercases all the column names.

* Define DataFrame utility configs

Define a set of utilities that can be applied to DataFrames during
loading or materialization. This is meant for simple operations that
occur specifically to loaded or materialized DataFrames, and should not
be used to replace business logic in solids.

* Add utilities to loader/materializer configs

Add a field per utility to the DataFrame loader/materializer configs.

* Move DataFrameUtilities to utils file

* Add set_index as a utility

* Map all method options

For the utilities that are direct mappings of DataFrame methods, adjust
the options to comprehensively match the method args and kwargs.

* Create tests for DataFrame utilities

* Wrap utility options in Field

* Apply utilities to loaded/materialized dataframes

Apply DataFrame utilities after loading and before materialization, if
specified.

* Refactor test and validate input with output

* Move reset_index before set_index

* Create canada.csv for testing utilities

This csv contains columns names with upper or mixed case letters and
has 16 rows of data to more easily test the utility functions.

* Read enabled flag

The sanitize_column_names config is defined as a flag. Take in the flag
value, and don't do anything if it is False.

* Test all utility functions individually

Test all the DataFrame utility functions individually using the
canada.csv data.

* Test a combination of all the utilities

* Rename sanitize_column_names to normalize_column_names

* Also change camel to snake case on normalize

Extend the normalize_column_names utility to also change camel case to
snake case.

* Revise description of normalize_column_names

* Fix indentation on sample options

* Refactor name normalize into a full function

The _normalize_names function can be used to normalize other names in
DataFrames. Split it out to a full function.

* Add drop to utils

The drop util can be used to drop columns from the DataFrame.

* Adjust code formatting with Black

* Apply Black code formatting changes

* Apply isort code formatting changes

* Add tests to ensure keys don't overlap