This adds a CLI to print the stored versions of each output produced by a pipeline. It can be executed by running `dagster pipeline print_versions <options>`.
Summary:
During the memoized development loop, it would be useful for users to see how their changes to a pipeline will change computed versions, and also to visualize which outputs need to be reconstructed. This CLI aims to provide a simplified way to do that.

In order to use it, a user would call `dagster pipeline print_versions <options>`, where the options are the same as those available from `dagster pipeline execute` (excluding solid selection and tags).

An example output:
{F325728}

A `None` address denotes that no address has been stored with the corresponding version.

Test Plan: Added tests in `test_memoized_development_cli.py`.

Reviewers: sandyryza, yuhan, alangenfeld

Reviewed By: sandyryza

Differential Revision: https://dagster.phacility.com/D4713