[dagit] dont error on partition page load
Summary: Would be nice to find a more clever solution to this problem when the repository context is not yet available so we return `''` which will fail in the query we use it for. This `skip` trick has the right product behavior - but its easy to miss when it should be used.

Test Plan: http://localhost:3000/pipeline/many_events/partitions - no more graphql error

Reviewers: dish, prha, bengotow, sashank

Reviewed By: dish

Differential Revision: https://dagster.phacility.com/D4711